requirements.txt
finance-datareader
pandas-ta-classic
python-telegram-bot
scipy
pandas
numpy
matplotlib
requests
